Remove redis-related tests, move tile-related tests to test_tile.py
